day 16

+++ src/main/scala/Day6.scala
@@ -13,8 +13,8 @@ object Day6 extends AoC:
   override def part2(lines: Vector[String]): Long =
     val time = NumRe.findAllIn(lines.head).mkString.toLong
     val dist = NumRe.findAllIn(lines(1)).mkString.toLong
-    val t0 = ((time - Math.sqrt(time * time - 4 * dist)) / 2).toLong
-    val t1 = ((time + Math.sqrt(time * time - 4 * dist)) / 2).toLong
+    val t0 = ((time - Math.sqrt(time.toDouble * time - 4 * dist)) / 2).toLong
+    val t1 = ((time + Math.sqrt(time.toDouble * time - 4 * dist)) / 2).toLong
     t1 - t0
   end part2
